debugging

+++ Noisy-BN/layers/noisy_bn.py
@@ -7,17 +7,21 @@ class NoisyBatchNormalization(tf.keras.layers.BatchNormalization):
 
     def __init__(self, alpha=0.01, *args, **kwargs):
         tf.keras.layers.BatchNormalization.__init__(self, *args, **kwargs)
-        self.alpha = np.float64(alpha)
+        self.alpha = alpha
 
     def call(self, inputs, training=None, alpha=None):
         if training is None:
             training = tf.keras.backend.learning_phase()
         input_shape = tf.keras.backend.int_shape(inputs)
         
+        # inputs = tf.cast(inputs, tf.float64)
         # Prepare broadcasting shape.
         ndim = len(input_shape)
         reduction_axes = list(range(len(input_shape)))
         del reduction_axes[self.axis[0]]
+
+        # print(reduction_axes)
+        # inputs = tf.cast(inputs, tf.float64)
 
 
         def broadcast_mean_var(x, ndim):
@@ -34,8 +38,8 @@ class NoisyBatchNormalization(tf.keras.layers.BatchNormalization):
 
         def mean_var_train():
             mean, var = tf.nn.moments(inputs,axes=reduction_axes)
-            self.moving_mean.assign(self.momentum * self.moving_mean + (1 - self.momentum) * mean)
-            self.moving_variance.assign(self.momentum * self.moving_variance + (1 - self.momentum) * var)
+            self.moving_mean.assign(self.momentum * self.moving_mean + (1. - self.momentum) * mean)
+            self.moving_variance.assign(self.momentum * self.moving_variance + (1. - self.momentum) * var)
             return [mean, var]
 
         mean, var = tf_utils.smart_cond(training, mean_var_train, mean_var_eval)
@@ -43,8 +47,11 @@ class NoisyBatchNormalization(tf.keras.layers.BatchNormalization):
 
         mean = broadcast_mean_var(mean, ndim)
         var = broadcast_mean_var(var, ndim)
+        # print(inputs, mean, var, self.epsilon)
+        # norm = tf.cast((inputs - mean), tf.float64) / tf.sqrt(tf.cast(var + self.epsilon, tf.float64))
         norm = (inputs - mean) / tf.sqrt(var + self.epsilon)
-        noise = tf.cast(tf.random.normal(tf.shape(inputs)), tf.float64)
+        noise = tf.random.normal(tf.shape(inputs))
+        # norm = tf.cast(norm, tf.float32)
         
         if alpha is None:
             norm_noise = tf_utils.smart_cond(
@@ -54,7 +61,7 @@ class NoisyBatchNormalization(tf.keras.layers.BatchNormalization):
                 )
         else:
 
-            norm_noise = norm + np.float64(alpha) * noise
+            norm_noise = norm + alpha * noise
 
-        return self.gamma * norm_noise + self.beta
+        return self.gamma * norm_noise + self.beta
